backend/requirements.txt
langchain-community
langchain-openai
python-dotenv
faiss-cpu
streamlit